% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TING-package.R
\name{pickAGE}
\alias{pickAGE}
\title{Build shiny APP that shows in each AGE group, the range of ACD that can cause
    the occurrence of MYOPIC.}
\usage{
pickAGE(select, input, output)
}
\value{
shiny APP shows that in each AGE group, the range of ACD that can cause
    the occurrence of MYOPIC.
}
\description{
Build shiny APP that shows in each AGE group, the range of ACD that can cause
    the occurrence of MYOPIC.
}
\examples{
pickAGE()
}
