% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TING-package.R
\name{Ting_correlation}
\alias{Ting_correlation}
\title{This function is used to do the pairwise partial correlations for
    independent variables and dependent variables. In addition, it gives us
    the p value as well as statistic for each correlation.}
\usage{
Ting_correlation(dataname, targetvariable)
}
\arguments{
\item{dataname}{The name of dataset}

\item{targetvariable}{One of independent variable}
}
\value{
p value from pairwise partial correlations
}
\description{
This function is used to do the pairwise partial correlations for
    independent variables and dependent variables. In addition, it gives us
    the p value as well as statistic for each correlation.
}
\examples{
Ting_correlation(mydata, "MYOPIC")
}
