% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TING-package.R
\name{Ting_graph}
\alias{Ting_graph}
\title{This function can build jitter plot between independent variables and
dependent variables.}
\usage{
Ting_graph(dataset, xname, targetvariable, title, subtitle, caption)
}
\arguments{
\item{dataset}{Dataset name}

\item{xname}{One of independent variables}

\item{targetvariable}{One of dependent variables}

\item{title}{Sitle of jitter plot}

\item{subtitle}{Subtitle of jitter plot}

\item{caption}{Caption of jitter plot}
}
\value{
Jitter plot
}
\description{
This function can build jitter plot between independent variables and
dependent variables.
}
\examples{
Ting_graph(mydata, AL, MYOPIC, "Use AL to estimate myopia", "R program",
"Produced by Ting Wang")
}
