% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TING-package.R
\name{Ting_regression}
\alias{Ting_regression}
\title{This function is used to get regression line on scatter plot for independent
    variables and dependent variables.}
\usage{
Ting_regression(dataname, xname, yname, xlab, ylab, title, caption)
}
\arguments{
\item{dataname}{Name of dataset}

\item{xname}{One of independent variables}

\item{yname}{One of dependent variables}

\item{xlab}{Label of x axis}

\item{ylab}{Label of y axis}

\item{title}{Title of plot}

\item{caption}{Caption of plot}
}
\value{
Scatter plot with regression line
}
\description{
This function is used to get regression line on scatter plot for independent
    variables and dependent variables.
}
\examples{
Ting_regression(mydata, GENDER, MYOPIC, "GENDER", "MYOPIC", "Regression
    GENDER versus MYOPIC", "Produced by Ting Wang")
}
