% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TING-package.R
\name{Ting_selection}
\alias{Ting_selection}
\title{This function is used to find significant correlations between independent
    variables and dependent variables.}
\usage{
Ting_selection(data.name)
}
\arguments{
\item{data.name}{Dataset name (dataset should include p vlaue from correlation
analysis)}
}
\value{
Significant pairwise partial correlations
}
\description{
This function is used to find significant correlations between independent
    variables and dependent variables.
}
\examples{
Ting_selection(data)
}
